+++ presets/bigCircles/src/index.ts
@@ -1,6 +1,6 @@
-import type { Engine } from "tsparticles-engine";
-import { loadBasic } from "tsparticles-basic";
-import { loadEmittersPlugin } from "tsparticles-plugin-emitters";
+import type { Engine } from "@tsparticles/engine";
+import { loadBasic } from "@tsparticles/basic";
+import { loadEmittersPlugin } from "@tsparticles/plugin-emitters";
 import { options } from "./options";
 
 /**