+++ presets/confetti/src/index.ts
@@ -1,13 +1,13 @@
-import type { Engine } from "tsparticles-engine";
-import { loadBasic } from "tsparticles-basic";
-import { loadEmittersPlugin } from "tsparticles-plugin-emitters";
-import { loadLifeUpdater } from "tsparticles-updater-life";
-import { loadMotionPlugin } from "tsparticles-plugin-motion";
-import { loadRollUpdater } from "tsparticles-updater-roll";
-import { loadRotateUpdater } from "tsparticles-updater-rotate";
-import { loadSquareShape } from "tsparticles-shape-square";
-import { loadTiltUpdater } from "tsparticles-updater-tilt";
-import { loadWobbleUpdater } from "tsparticles-updater-wobble";
+import type { Engine } from "@tsparticles/engine";
+import { loadBasic } from "@tsparticles/basic";
+import { loadEmittersPlugin } from "@tsparticles/plugin-emitters";
+import { loadLifeUpdater } from "@tsparticles/updater-life";
+import { loadMotionPlugin } from "@tsparticles/plugin-motion";
+import { loadRollUpdater } from "@tsparticles/updater-roll";
+import { loadRotateUpdater } from "@tsparticles/updater-rotate";
+import { loadSquareShape } from "@tsparticles/shape-square";
+import { loadTiltUpdater } from "@tsparticles/updater-tilt";
+import { loadWobbleUpdater } from "@tsparticles/updater-wobble";
 import { options } from "./options";
 
 /**