+++ presets/fire/src/options.ts
@@ -1,4 +1,4 @@
-import { ClickMode, type ISourceOptions } from "tsparticles-engine";
+import { ClickMode, type ISourceOptions } from "@tsparticles/engine";
 
 export const options: ISourceOptions = {
     fpsLimit: 40,