+++ presets/firefly/src/options.ts
@@ -1,4 +1,4 @@
-import { HoverMode, type ISourceOptions } from "tsparticles-engine";
+import { HoverMode, type ISourceOptions } from "@tsparticles/engine";
 
 export const options: ISourceOptions = {
     fullScreen: {