+++ presets/fireworks/src/index.ts
@@ -1,12 +1,12 @@
-import type { Engine } from "tsparticles-engine";
-import { loadBasic } from "tsparticles-basic";
-import { loadDestroyUpdater } from "tsparticles-updater-destroy";
-import { loadEmittersPlugin } from "tsparticles-plugin-emitters";
-import { loadLifeUpdater } from "tsparticles-updater-life";
-import { loadLineShape } from "tsparticles-shape-line";
-import { loadRotateUpdater } from "tsparticles-updater-rotate";
-import { loadSoundsPlugin } from "tsparticles-plugin-sounds";
-import { loadStrokeColorUpdater } from "tsparticles-updater-stroke-color";
+import type { Engine } from "@tsparticles/engine";
+import { loadBasic } from "@tsparticles/basic";
+import { loadDestroyUpdater } from "@tsparticles/updater-destroy";
+import { loadEmittersPlugin } from "@tsparticles/plugin-emitters";
+import { loadLifeUpdater } from "@tsparticles/updater-life";
+import { loadLineShape } from "@tsparticles/shape-line";
+import { loadRotateUpdater } from "@tsparticles/updater-rotate";
+import { loadSoundsPlugin } from "@tsparticles/plugin-sounds";
+import { loadStrokeColorUpdater } from "@tsparticles/updater-stroke-color";
 import { options } from "./options";
 
 /**