+++ presets/fountain/src/index.ts
@@ -1,7 +1,7 @@
-import type { Engine } from "tsparticles-engine";
-import { loadBasic } from "tsparticles-basic";
-import { loadDestroyUpdater } from "tsparticles-updater-destroy";
-import { loadEmittersPlugin } from "tsparticles-plugin-emitters";
+import type { Engine } from "@tsparticles/engine";
+import { loadBasic } from "@tsparticles/basic";
+import { loadDestroyUpdater } from "@tsparticles/updater-destroy";
+import { loadEmittersPlugin } from "@tsparticles/plugin-emitters";
 import { options } from "./options";
 
 /**