+++ presets/fountain/src/options.ts
@@ -1,4 +1,4 @@
-import { type ISourceOptions, MoveDirection, OutMode } from "tsparticles-engine";
+import { type ISourceOptions, MoveDirection, OutMode } from "@tsparticles/engine";
 
 export const options: ISourceOptions = {
     fpsLimit: 120,
@@ -64,7 +64,9 @@ export const options: ISourceOptions = {
             },
             trail: {
                 enable: true,
-                fillColor: "#fff",
+                fill: {
+                    color: "#fff",
+                },
                 length: 3,
             },
         },