+++ presets/seaAnemone/src/options.ts
@@ -1,5 +1,5 @@
-import type { ISourceOptions } from "tsparticles-engine";
-import { curvesPathName } from "tsparticles-path-curves";
+import type { ISourceOptions } from "@tsparticles/engine";
+import { curvesPathName } from "@tsparticles/path-curves";
 
 export const presetName = "seaAnemone";
 
@@ -27,7 +27,9 @@ export const options: ISourceOptions = {
             speed: 2,
             straight: false,
             trail: {
-                fillColor: "#000",
+                fill: {
+                    color: "#000",
+                },
                 length: 30,
                 enable: true,
             },