+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [2.12.0](https://github.com/tsparticles/presets/compare/v2.11.0...v2.12.0) (2023-08-04)
+# [3.0.0-beta.1](https://github.com/tsparticles/presets/compare/v2.11.0...v3.0.0-beta.1) (2023-08-04)
 
 
 ### Features
@@ -16,7 +16,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [2.11.0](https://github.com/tsparticles/presets/compare/v2.10.1...v2.11.0) (2023-07-13)
 
-**Note:** Version bump only for package tsparticles-presets-workspace
+**Note:** Version bump only for package @tsparticles/presets-workspace
 
 
 
@@ -24,7 +24,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [2.10.1](https://github.com/tsparticles/presets/compare/v2.10.0...v2.10.1) (2023-06-04)
 
-**Note:** Version bump only for package tsparticles-presets-workspace
+**Note:** Version bump only for package @tsparticles/presets-workspace