+++ presets/hyperspace/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [2.12.0](https://github.com/tsparticles/presets/compare/v2.11.0...v2.12.0) (2023-08-04)
+# [3.0.0-beta.1](https://github.com/tsparticles/presets/compare/v2.11.0...v3.0.0-beta.1) (2023-08-04)
 
 
 ### Features